Add PQ presets (FALCON1024, SPHINCS+), fix all 369 rung tests, regenerate tx docs
- Add FALCON-1024 and SPHINCS+-SHA2 presets to test framework, engine, and doc generator
- Increase proxy size limits to 256KB for SPHINCS+ witnesses (~50KB)
- Make decode-tx non-fatal for oversized PQ transactions
- Write 10 new gap coverage tests: key size mismatch, empty pubkey, multisig
  commit mismatch, DoS-safe recursion depth, relay limit, mixed PQ+Schnorr
  carry-forward, PQ scheme round-trip, coil UNLOCK_TO/COVENANT round-trip
- Fix 7 pre-existing test bugs: witness size padding (PREIMAGE 252B max),
  relay error string, 3 context-dependent blocks (RECURSE_MODIFIED, COSIGN,
  HYSTERESIS_FEE) expecting SATISFIED without tx context, TIMELOCKED_MULTISIG
  missing second PUBKEY
- Regenerate all 36 tx docs with real signet data and engine-matched colors
- Update docs index to 8 categories / 36 entries with family color coding
- Mark all BIP readiness audit items resolved

+++ src/rung/evaluator.cpp
@@ -104,7 +104,7 @@ static int64_t ReadNumeric(const RungField& field)
     for (size_t i = 0; i < field.data.size(); ++i) {
         val |= static_cast<uint32_t>(field.data[i]) << (8 * i);
     }
-    return static_cast<int64_t>(val);
+    return static_cast<int64_t>(static_cast<int32_t>(val));
 }
 
 /** Helper: check if a pubkey field exists and has valid size.
